add doctor command

+++ run.sh
@@ -1,6 +1,6 @@
 #! /bin/bash -e
 
-source .zprofile
+[[ -f "~/.zprofile" ]] && source .zprofile
 source ./run.conf
 
 if ! $GENESIS_RUN; then
@@ -11,23 +11,33 @@ fi
 
 if ! [ -z $GENESIS_PROFILE ]; then
     # NOTE: these aliases will not work unless you source them in the .zshrc file
+    echo "setting profile: $GENESIS_PROFILE"
 
     if ! [ -z $GITHUB_USER ]; then
         # TODO: should with work with other peoples .zprofile?
+        echo "setting github user: $GITHUB_USER"
 
         sed -i "s/export GITHUB_USER=.*/export GITHUB_USER=$GITHUB_USER/" $GENESIS_PROFILE
     fi
 
     # make the aliases directory
-    rm ~/.zprofile
+    [[ -f "~/.zprofile" ]] && rm ~/.zprofile
 
     # Add in the aliases
     cp $GENESIS_PROFILE ~/.zprofile
+    source ~/.zprofile
 fi
 
 # Add in the .zshrc file
-rm ~/.zshrc
+[[ -f "~/.zshrc" ]] && rm ~/.zshrc
+
 cp ./.zshrc ~/.zshrc
+
+echo $SHELL
+
+if ! [ -x "$(command -v lsb_release)" ]; then
+   sudo apt install lsb-core
+fi
 
 release=`lsb_release -i`
